Combine Add / Remove Treasury Cw20s actions
