+++ packages/i18n/locales/en/translation.json
@@ -440,6 +440,7 @@
     "manageStorageItemsDescription": "Manage Storage Items associated with your DAO. Set or remove key / value pairs.",
     "manageSubDaosActionDescription": "Recognize or remove SubDAOs from the DAO.",
     "manageTreasuryNftsDescription": "Manage NFT Collections displayed on your DAO's treasury page.",
+    "manageTreasuryTokensDescription": "Manage tokens displayed on your DAO's treasury page.",
     "members": "members",
     "membershipDaoNotMemberInfo_dao": "You are not a member of {{daoName}}. To become a member, existing members must vote to add you.",
     "membershipDaoNotMemberInfo_proposal": "You were not a member of {{daoName}} when this proposal was created, so you cannot vote on it.",
@@ -705,6 +706,7 @@
     "manageStaking": "Manage staking",
     "manageSubDaos": "Manage SubDAOs",
     "manageTreasuryNfts": "Manage Treasury NFTs",
+    "manageTreasuryTokens": "Manage Treasury Tokens",
     "memberOfDao": "Member of {{daoName}}",
     "members": "Members",
     "membership": "Membership",
