+++ packages/stateful/actions/actions/index.tsx
@@ -1,10 +1,10 @@
 import { Action, ActionOptions } from '@dao-dao/types/actions'
 import { DISABLED_ACTIONS } from '@dao-dao/utils'
 
-import { makeAddCw20Action } from './AddCw20'
 import { makeCustomAction } from './Custom'
 import { makeExecuteAction } from './Execute'
 import { makeInstantiateAction } from './Instantiate'
+import { makeManageCw20Action } from './ManageCw20'
 import { makeManageCw721Action } from './ManageCw721'
 import { makeManageSubDaosAction } from './ManageSubDaos'
 import { makeMigrateAction } from './MigrateContract'
@@ -13,7 +13,6 @@ import {
   makeMintNftAction,
   makeTransferNftAction,
 } from './nft'
-import { makeRemoveCw20Action } from './RemoveCw20'
 import { makeRemoveItemAction } from './RemoveItem'
 import { makeSetItemAction } from './SetItem'
 import { makeSpendAction } from './Spend'
@@ -31,8 +30,7 @@ export const getActions = (options: ActionOptions): Action[] => {
     makeSpendAction,
     makeStakeAction,
     makeUpdateInfoAction,
-    makeAddCw20Action,
-    makeRemoveCw20Action,
+    makeManageCw20Action,
     makeManageCw721Action,
     makeMintNftAction,
     makeTransferNftAction,
