+++ packages/stateful/actions/components/ManageCw20.stories.tsx
@@ -2,16 +2,16 @@ import { ComponentMeta, ComponentStory } from '@storybook/react'
 
 import { ReactHookFormDecorator } from '@dao-dao/storybook'
 
-import { AddCw20Component } from './AddCw20'
+import { ManageCw20Component } from './ManageCw20'
 
 export default {
-  title: 'DAO DAO / packages / stateful / actions / components / AddCw20',
-  component: AddCw20Component,
+  title: 'DAO DAO / packages / stateful / actions / components / ManageCw20',
+  component: ManageCw20Component,
   decorators: [ReactHookFormDecorator],
-} as ComponentMeta<typeof AddCw20Component>
+} as ComponentMeta<typeof ManageCw20Component>
 
-const Template: ComponentStory<typeof AddCw20Component> = (args) => (
-  <AddCw20Component {...args} />
+const Template: ComponentStory<typeof ManageCw20Component> = (args) => (
+  <ManageCw20Component {...args} />
 )
 
 export const Default = Template.bind({})
