+++ packages/stateful/actions/components/index.tsx
@@ -1,13 +1,12 @@
 export * from './ActionCard'
 export * from './ActionsRenderer'
-export * from './AddCw20'
 export * from './Custom'
 export * from './Execute'
 export * from './Instantiate'
+export * from './ManageCw20'
 export * from './ManageCw721'
 export * from './MigrateContract'
 export * from './NativeCoinSelector'
-export * from './RemoveCw20'
 export * from './Spend'
 export * from './StakingActions'
 export * from './UpdateAdmin'
